refactor: move services, providers and resolvers to shared/data-access
