+++ apps/iswavle/project.json
@@ -69,13 +69,7 @@
         "development": {
           "optimization": false,
           "extractLicenses": false,
-          "sourceMap": true,
-          "fileReplacements": [
-            {
-              "replace": "apps/iswavle/src/environments/environment.ts",
-              "with": "apps/iswavle/src/environments/environment.development.ts"
-            }
-          ]
+          "sourceMap": true
         }
       },
       "defaultConfiguration": "production",
