+++ apps/iswavle/src/app/features/docs/ui/doc-toc/doc-toc.component.ts
@@ -14,7 +14,7 @@ import { TocItem } from '@iswavle/shared/utils';
 import { NzDividerModule } from 'ng-zorro-antd/divider';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { NgTemplateOutlet } from '@angular/common';
-import { ENVIRONMENT } from '@app-shared/providers/environment';
+import { ENVIRONMENT } from '@iswavle/shared/data-access';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 
 @Component({
