+++ apps/iswavle/src/app/features/exercises/exercises-list/exercises.component.ts
@@ -12,7 +12,7 @@ import { ActivatedRoute, Router, RouterLink } from '@angular/router';
 import { ExercisesNavigation, ExercisesTableData } from '@iswavle/shared/utils';
 import { ExerciseDifficultyPipe, EncodeURIPipe } from '@iswavle/shared/ui';
 import { EXERCISE_TAG_PATH_MAP, DIFFICULTY_TEXT } from '@iswavle/shared/utils';
-import { ExercisesService, LayoutService } from '@app-shared/services';
+import { ExercisesService, LayoutService } from '@iswavle/shared/data-access';
 import { map } from 'rxjs/operators';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { NzLayoutModule } from 'ng-zorro-antd/layout';
