+++ libs/shared/data-access/project.json
@@ -14,6 +14,16 @@
         "tsConfig": "libs/shared/data-access/tsconfig.spec.json"
       }
     },
+    "configurations": {
+      "development": {
+        "fileReplacements": [
+          {
+            "replace": "libs/shared/data-access/src/lib/environments/environment.ts",
+            "with": "libs/shared/data-access/src/lib/environments/environment.development.ts"
+          }
+        ]
+      }
+    },
     "lint": {
       "executor": "@nx/eslint:lint"
     }
