--- libs/shared/data-access/src/lib/data-access/data-access.html
@@ -1,1 +1,0 @@
-<p>DataAccess works!</p>
