--- libs/shared/data-access/src/lib/data-access/data-access.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { DataAccess } from './data-access';
-
-describe('DataAccess', () => {
-  let component: DataAccess;
-  let fixture: ComponentFixture<DataAccess>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [DataAccess],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(DataAccess);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
