--- libs/shared/data-access/src/lib/data-access/data-access.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'sw-data-access',
-  imports: [],
-  templateUrl: './data-access.html',
-  styleUrl: './data-access.less',
-})
-export class DataAccess {}
