+++ libs/shared/data-access/src/lib/providers/code-loader.ts
@@ -1,6 +1,6 @@
 import { InjectionToken } from '@angular/core';
 import { NzTreeNodeOptions } from 'ng-zorro-antd/tree';
-import { CodeLoaderService } from '@app-shared/services';
+import { CodeLoaderService } from '../services';
 import { ContentLoader } from '@iswavle/shared/utils';
 
 export const CODE_LOADER = new InjectionToken<ContentLoader<NzTreeNodeOptions>>(
