+++ libs/shared/data-access/src/lib/providers/docs-content-loader.ts
@@ -1,7 +1,7 @@
 import { InjectionToken } from '@angular/core';
 
 import { ContentLoader, DocContent } from '@iswavle/shared/utils';
-import { ContentLoaderService } from '@app-shared/services';
+import { ContentLoaderService } from '../services';
 
 export const DOCS_CONTENT_LOADER = new InjectionToken<
   ContentLoader<DocContent>
