+++ libs/shared/data-access/src/lib/providers/environment.ts
@@ -1,6 +1,6 @@
 import { InjectionToken } from '@angular/core';
 import { Environment } from '@iswavle/shared/utils';
-import { environment } from '../../../environments/environment';
+import { environment } from '../environments/environment';
 
 export const ENVIRONMENT = new InjectionToken<Environment>('ENVIRONMENT', {
   providedIn: 'root',
