+++ libs/shared/data-access/src/lib/providers/exercises-map.ts
@@ -1,6 +1,6 @@
 import { InjectionToken } from '@angular/core';
 import { ContentLoader, ExercisesNavigation } from '@iswavle/shared/utils';
-import { ExercisesMapService } from '@app-shared/services';
+import { ExercisesMapService } from '../services';
 
 export const EXERCISES_MAP = new InjectionToken<
   ContentLoader<ExercisesNavigation[]>
