+++ libs/shared/data-access/src/lib/providers/webcontainer-state.ts
@@ -1,6 +1,6 @@
 import { InjectionToken, Provider } from '@angular/core';
 import { WebContainerState } from '@iswavle/shared/utils';
-import { WebContainerService } from '@app-shared/services';
+import { WebContainerService } from '../services';
 
 export const WEBCONTAINER_STATE = new InjectionToken<WebContainerState>(
   'WEBCONTAINER_STATE',
