+++ libs/shared/data-access/src/lib/resolvers/exercises-map.resolver.ts
@@ -1,7 +1,7 @@
 import { inject } from '@angular/core';
 import { ActivatedRouteSnapshot, ResolveFn } from '@angular/router';
 import { Params, ExercisesNavigation } from '@iswavle/shared/utils';
-import { EXERCISES_MAP } from '@app-shared/providers';
+import { EXERCISES_MAP } from '../providers';
 
 export const exercisesMapResolver: ResolveFn<ExercisesNavigation[] | null> = (
   route: ActivatedRouteSnapshot,
