+++ libs/shared/data-access/src/lib/resolvers/exercises.resolver.ts
@@ -1,6 +1,6 @@
 import { inject } from '@angular/core';
 import { ResolveFn, ActivatedRouteSnapshot } from '@angular/router';
-import { EXERCISES_LOADER } from '@app-shared/providers';
+import { EXERCISES_LOADER } from '../providers';
 import { Params, ExercisesContent } from '@iswavle/shared/utils';
 
 export const exercisesResolver: ResolveFn<ExercisesContent | null> = (
