+++ libs/shared/data-access/src/lib/services/article.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, inject } from '@angular/core';
 import { NavigationEnd, Router } from '@angular/router';
 import { NavigationTreeNode } from '@iswavle/shared/utils';
-import { DOC_NAVIGATION } from '@app-shared/providers';
+import { DOC_NAVIGATION } from '../providers';
 import { filter, map } from 'rxjs';
 
 @Injectable({
