+++ libs/shared/data-access/src/lib/services/index.ts
@@ -4,7 +4,6 @@ export * from './code-loader.service';
 export * from './theme.service';
 export * from './meta.service';
 export * from './title.strategy';
-export * from './search.service';
 export * from './article.service';
 export * from './contributors.service';
 export * from './layout.service';
